Tame the matrix editing to one at a time
Adds a select list above the matrix table, which allows one of the items to be opened for editing. In reality it just shows/hides a row of the table, so you can still make changes to multiple matrices before Saving them all.

The currently edited matrix is stored in localStorage when selecting via the dropdown. Note it does *not* automatically select any new matrices that are created. This enables rapid addition of multiple new matrices.